修改了LDPC编解码算法结构

+++ src/test.h
@@ -1,6 +1,6 @@
 #ifndef TEST_H
 #define TEST_H
 
-void run(int frames, int snr_db, int iteration, float alpha, float beta);
+void run(int frames, double snr_db, int iteration, float alpha, float beta);
 
 #endif
